Respond to peer review
